Rainfall Calculator
After entering the number of years and this program will ask for the number of inches of rainfall per month. It will calculate yearly totals and monthly averages for each year. At the end of the program, it will display the total months, the total rainfall for all the years in inches, as well as the average rainfall per month in inches.