Country Capital Quiz
This program uses an array of countries and capitals to quiz you on which countries have corresponding capitals. It keeps track of how many answers you have gotten right or wrong, and then displays this information at the end of the quiz. The user can end the quiz whenever they want by selecting not to continue after a quiz question.