Share Profitablility Comparison
This program asks the user to input the name of the stock, the number of purchases for that stock, the number of shares purchased each time it was purchased, the price per share of the stock at the time it was purchased, and finally today's price per share of the stock. After this information is entered, the program displays the net gain/loss of the stock. The program asks for the information above again for the second stock, and compares the two stocks to find out which is more profitable and displays that information.